+++ src/engine/arcan_lua.c
@@ -2288,10 +2288,18 @@ static int imageresizestorage(lua_State* ctx)
 	struct rendertarget* rtgt = arcan_vint_findrt(vobj);
 	if (rtgt){
 		agp_resize_rendertarget(rtgt->art, w, h);
+
+		ssize_t view_w = luaL_optnumber(ctx, 4, w);
+		ssize_t view_h = luaL_optnumber(ctx, 5, h);
+		ssize_t x = luaL_optnumber(ctx, 6, 0);
+		ssize_t y = luaL_optnumber(ctx, 7, 0);
+
 		if (!rtgt->inv_y)
-			build_orthographic_matrix(rtgt->projection, 0, w, 0, h, 0, 1);
+			build_orthographic_matrix(rtgt->projection, x, w, y, h, 0, 1);
 		else
-			build_orthographic_matrix(rtgt->projection, 0, w, h, 0, 0, 1);
+			build_orthographic_matrix(rtgt->projection, x, w, h, y, 0, 1);
+
+		agp_rendertarget_viewport(rtgt->art, x, y, x+view_w, y+view_h);
 	}
 	else
 		arcan_video_resizefeed(id, w, h);